requirements.txt
attrs==22.2.0
basemap @ file:///C:/Users/winzigt/PycharmProjects/svg/venv/Lib/site-packages/basemap-1.3.6-cp38-cp38-win32.whl
basemap-data==1.3.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Fiona @ file:///C:/Users/winzigt/PycharmProjects/svg/venv/Lib/site-packages/Fiona-1.8.22-cp38-cp38-win32.whl
fonttools==4.38.0
GDAL @ file:///C:/Users/winzigt/PycharmProjects/svg/venv/Lib/site-packages/GDAL-3.6.2-cp38-cp38-win32.whl
geopandas==0.12.2
idna==3.4
importlib-metadata==6.0.0
kiwisolver==1.4.4
llvmlite==0.39.1
lxml==4.9.2
matplotlib==3.6.3
munch==2.5.0
numba==0.56.4
numpy==1.23.5
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pyparsing==3.0.9
pyproj==3.4.1
pyshp==2.3.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
scipy==1.9.1
shapely==2.0.0
shapley==1.0.3
six==1.16.0
svg.path==6.2
svgpathtools==1.5.1
svgutils==0.3.4
svgwrite==1.4.3
urllib3==1.26.14
zipp==3.11.0
